fuzz fix
request-checks: true
